+++ app/src/main/java/com/trianguloy/urlchecker/activities/MainActivity.java
@@ -90,7 +90,7 @@ public class MainActivity extends Activity {
     }
 
     public void aboutToast(View view) {
-        Toast.makeText(this, getString(R.string.app_name) + " - " + getString(R.string.trianguloy), Toast.LENGTH_SHORT).show();
+        Toast.makeText(this, getString(R.string.app_name) + " - " + getString(R.string.shiroikuma), Toast.LENGTH_SHORT).show();
     }
 
 }